✨ feat: add search input component; update various interfaces to use string IDs; enhance cart service with user ID handling; improve styling for plans and search company dialog

+++ src/app/pages/categories/components/categories/categories.component.ts
@@ -8,6 +8,9 @@ import { AuthService } from 'src/app/domain/auth/services/auth.service';
 import { LoadingService } from '@shared/services/loading/loading.service';
 import { LoadingComponent } from '@shared/components/loading/loading.component';
 import { ToastrService } from 'ngx-toastr';
+import { ConfirmDialogComponent } from '@shared/dialogs/confirm-dialog/confirm-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { FoodService } from '@shared/services/food/food.service';
 
 @Component({
   selector: 'app-categories',
@@ -22,16 +25,19 @@ export class CategoriesComponent {
 
   private localStorageService = inject(LocalStorageService);
   private categoryService = inject(CategoryService);
+  private foodService = inject(FoodService);
   private authService = inject(AuthService);
   private toastr = inject(ToastrService);
+  private dialog = inject(MatDialog);
+
 
   public loadingService = inject(LoadingService);
   public isLoading = signal<boolean>(false);
   public isAdmin = this.authService.isAdmin;
   public existingAssociation!: boolean;
-  public isCategoryAssociatedMap: { [key: number]: boolean } = {};
+  public isCategoryAssociatedMap: Record<string, boolean> = {};
 
-  public companyId = this.localStorageService.getSignal('companyId', 0);
+  public companyId = this.localStorageService.getSignal('companyId', '0');
   public companyName = this.localStorageService.getSignal<string>(
     'companyName',
     '[]'
@@ -41,7 +47,7 @@ export class CategoriesComponent {
     this.getCategoryAssociations();
   }
 
-  async addCategoryToCompany(categoryId: number) {
+  async addCategoryToCompany(categoryId: string) {
     this.loadingService.showLoading();
     try {
       const data = {
@@ -50,6 +56,7 @@ export class CategoriesComponent {
       }
 
       await this.categoryService.insert('company_categories', data);
+
       this.isCategoryAssociatedMap[categoryId] = true;
 
       this.toastr.success("Categoria foi adicionada com sucesso")
@@ -58,21 +65,41 @@ export class CategoriesComponent {
     }
   }
 
-  async removeCategoryFromCompany(categoryId: number) {
-    this.loadingService.showLoading();
-    try {
+  async removeCategoryFromCompany(categoryId: string) {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '350px',
+      data: {
+        title: 'Excluir Categoria',
+        message: 'Tem certeza que deseja excluir essa categoria?',
+        confirmText: 'Excluir',
+        cancelText: 'Cancelar',
+        loading: this.loadingService.loading(),
+      },
+    });
 
-      await this.categoryService.deleteByFilter('company_categories', {
-        company_id: this.companyId(),
-        category_id: categoryId,
-      });
+    dialogRef.afterClosed().subscribe(async (result) => {
+      if (result) {
+        try {
+          const foods = await this.foodService.getFoodsByCategory(categoryId, this.companyId());
 
-      delete this.isCategoryAssociatedMap[categoryId];
+          if (foods && foods.length > 0) {
+            this.toastr.warning('Não é possível excluir a categoria pois há comidas associadas a ela.');
+            return;
+          }
+          
+          await this.categoryService.deleteByFilter('company_categories', {
+            company_id: this.companyId(),
+            category_id: categoryId,
+          });
 
-      this.toastr.success('Categoria removida com sucesso!');
-    } finally {
-      this.loadingService.hideLoading();
-    }
+          delete this.isCategoryAssociatedMap[categoryId];
+
+          this.toastr.success('Categoria removida com sucesso!');
+        } finally {
+          this.loadingService.hideLoading();
+        }
+      }
+    });
   }
 
   private async getCategoryAssociations(): Promise<void> {
@@ -90,6 +117,6 @@ export class CategoriesComponent {
   }
 
   public trackById(index: number, item: iCategory): number {
-    return item.id;
+    return +item.id;
   }
 }
